# app/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Lukon API",
    "description": "API for managing Lukons",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/create": {
      "post": {
        "summary": "Create a new Lukon",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Lukon"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lukon created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/search": {
      "get": {
        "summary": "Search Lukons",
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "schema": {
              "type": "string"
            },
            "description": "Keyword to search for"
          },
          {
            "in": "query",
            "name": "tags",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated list of tags to filter by"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Lukon"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Lukon": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "problem": {
            "type": "string"
          },
          "solution": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "description",
          "problem",
          "solution",
          "user_id"
        ]
      }
    }
  }
}